requirements.txt
Django==4.0.3
django-phonenumber-field
phonenumbers
twilio
# for Docker/deployment
mysqlclient
uwsgi
python-dotenv
